refact: moving folder to domain srp
