+++ src/test/java/xpian94/IntegrationTestBase.java
@@ -1,4 +1,4 @@
-package xpian94.todo;
+package xpian94;
 
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
