+++ src/test/java/xpian94/IntegrationTestBaseWithoutDataSource.java
@@ -1,4 +1,4 @@
-package xpian94.todo;
+package xpian94;
 
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
