+++ src/test/java/xpian94/home/HttpRequestSanityCheckTest.java
@@ -1,9 +1,11 @@
-package xpian94.todo;
+package xpian94.home;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.web.client.TestRestTemplate;
+import xpian94.IntegrationTestBaseWithoutDataSource;
+import xpian94.todo.TodoService;
 
 import static org.assertj.core.api.Assertions.assertThat;
